rename action modules
