+++ src/js/config/default.json
@@ -1,4 +1,6 @@
 {
   "rootSelector": "#game-root-container",
-  "startingLevel": "one"
+  "startingLevel": "one",
+  "startingArmor": "",
+  "startingWeapon": ""
 }
